Prepare Dockerfile for experimental registry
For the registry running on epr-experimental.elastic.co we need to have a different set of packages. Because of this I created an alpha branch where we can pull in the packages we need.

This is a first iteration on it to at least have the Dockerimage ready in a separate branch.